move dialogbuttons and lightbulb back. clean up references
