+++ apps/src/authoredHints.js
@@ -5,11 +5,11 @@
 
 var dom = require('./dom');
 var msg = require('./locale');
-var HintSelect = require('./templates/HintSelect.jsx');
-var HintsDisplay = require('./templates/HintsDisplay.jsx');
-var HintDialogContent = require('./templates/HintDialogContent.jsx');
+var HintSelect = require('./templates/instructions/HintSelect.jsx');
+var HintsDisplay = require('./templates/instructions/HintsDisplay.jsx');
+var HintDialogContent = require('./templates/instructions/HintDialogContent.jsx');
 var authoredHintUtils = require('./authoredHintUtils');
-var Lightbulb = require('./templates/instructions/Lightbulb.jsx');
+var Lightbulb = require('./templates/Lightbulb.jsx');
 
 var AuthoredHints = function (studioApp) {
   this.studioApp_ = studioApp;
