+++ apps/src/templates/instructions/HintDialogContent.jsx
@@ -1,6 +1,6 @@
 var msg = require('../../locale');
-var Lightbulb = require('./Lightbulb.jsx');
-var ReadOnlyBlockSpace = require('./ReadOnlyBlockSpace.jsx');
+var Lightbulb = require('../Lightbulb.jsx');
+var ReadOnlyBlockSpace = require('../ReadOnlyBlockSpace.jsx');
 
 var HintDialogContent = React.createClass({
   propTypes: {
